Proper sizes data type

+++ GenTreeBillboardTexture.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "GenTreeBillboardTexture.h"
 
-void GenTreeBillboardTexture(char *data, int width, int height, unsigned int seed) {
+void GenTreeBillboardTexture(char *data, size_t width, size_t height, unsigned int seed) {
 	TreeNode* tree = GenTreeBillboardTexture_grammar("tree1", 40, seed);  // args: name, iteration, seed
  	tree->printTree();
